# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISP Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all service plans
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────────
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Cancel/delete a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── INVOICES ──────────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: invoiceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── TICKETS ───────────────────────────────────────────────────────────────────
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all support tickets
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tickets/{id}/reply:
    post:
      operationId: replyToTicket
      tags: [tickets]
      summary: Add a reply to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketReply"
  /tickets/{id}/replies:
    get:
      operationId: getTicketReplies
      tags: [tickets]
      summary: Get replies for a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketReply"

  # ── EQUIPMENT ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - address
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Plan:
      type: object
      required:
        - id
        - name
        - downloadSpeed
        - uploadSpeed
        - price
        - billingCycle
        - isActive
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        downloadSpeed:
          type: integer
          description: Download speed in Mbps
        uploadSpeed:
          type: integer
          description: Upload speed in Mbps
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, annually]
        isActive:
          type: boolean
        createdAt:
          type: string
    PlanInput:
      type: object
      required: [name, downloadSpeed, uploadSpeed, price, billingCycle]
      properties:
        name:
          type: string
        description:
          type: string
        downloadSpeed:
          type: integer
        uploadSpeed:
          type: integer
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, annually]
        isActive:
          type: boolean
    PlanUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        downloadSpeed:
          type: integer
        uploadSpeed:
          type: integer
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, annually]
        isActive:
          type: boolean

    # ── SUBSCRIPTIONS ──────────────────────────────────────────────────────────
    Subscription:
      type: object
      required:
        - id
        - customerId
        - planId
        - status
        - startDate
        - createdAt
      properties:
        id:
          type: integer
        customerId:
          type: integer
        planId:
          type: integer
        status:
          type: string
          enum: [active, suspended, cancelled, expired]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        macAddress:
          type: ["string", "null"]
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        plan:
          $ref: "#/components/schemas/Plan"
    SubscriptionInput:
      type: object
      required: [customerId, planId, startDate]
      properties:
        customerId:
          type: integer
        planId:
          type: integer
        status:
          type: string
          enum: [active, suspended, cancelled, expired]
        startDate:
          type: string
        endDate:
          type: string
        ipAddress:
          type: string
        macAddress:
          type: string
    SubscriptionUpdate:
      type: object
      properties:
        planId:
          type: integer
        status:
          type: string
          enum: [active, suspended, cancelled, expired]
        endDate:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        macAddress:
          type: ["string", "null"]

    # ── INVOICES ───────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required:
        - id
        - customerId
        - invoiceId
        - amount
        - method
        - status
        - createdAt
      properties:
        id:
          type: integer
        customerId:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, bank_transfer, mobile_money, card, cheque]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
    PaymentInput:
      type: object
      required: [customerId, invoiceId, amount, method]
      properties:
        customerId:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, bank_transfer, mobile_money, card, cheque]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        reference:
          type: string
        notes:
          type: string

    # ── TICKETS ────────────────────────────────────────────────────────────────
    Ticket:
      type: object
      required:
        - id
        - customerId
        - subject
        - description
        - status
        - priority
        - createdAt
      properties:
        id:
          type: integer
        customerId:
          type: integer
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
    TicketInput:
      type: object
      required: [customerId, subject, description]
      properties:
        customerId:
          type: integer
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
        assignedTo:
          type: string
    TicketUpdate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
    TicketReply:
      type: object
      required: [id, ticketId, message, author, isStaff, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        message:
          type: string
        author:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
    TicketReplyInput:
      type: object
      required: [message, author, isStaff]
      properties:
        message:
          type: string
        author:
          type: string
        isStaff:
          type: boolean

    # ── EQUIPMENT ──────────────────────────────────────────────────────────────
...