# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanBillingCycle } from './planBillingCycle';

export interface Plan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Download speed in Mbps */
  downloadSpeed: number;
  /** Upload speed in Mbps */
  uploadSpeed: number;
  price: number;
  billingCycle: PlanBillingCycle;
  isActive: boolean;
  createdAt: string;
}