# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanUpdateBillingCycle } from './planUpdateBillingCycle';

export interface PlanUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  downloadSpeed?: number;
  uploadSpeed?: number;
  price?: number;
  billingCycle?: PlanUpdateBillingCycle;
  isActive?: boolean;
}