# lib/api-zod/src/generated/types/subscriptionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionUpdateStatus } from './subscriptionUpdateStatus';

export interface SubscriptionUpdate {
  planId?: number;
  status?: SubscriptionUpdateStatus;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  macAddress?: string | null;
}